@@ -2,9 +2,9 @@ __package__ = 'archivebox.core'
 
 from django import forms
 
-from ..util import URL_REGEX
+from archivebox.misc.util import URL_REGEX
 from ..parsers import PARSERS
-from ..vendor.taggit_utils import edit_string_for_tags, parse_tags
+from taggit.utils import edit_string_for_tags, parse_tags
 
 PARSER_CHOICES = [
     (parser_key, parser[0])
