@@ -1,13 +1,17 @@
 # Generated by Django 3.0.8 on 2020-11-04 12:25
 
+import os
 import json
 from pathlib import Path
 
 from django.db import migrations, models
 import django.db.models.deletion
 
-from config import CONFIG
 from index.json import to_json
+
+DATA_DIR = Path(os.getcwd()).resolve()                    # archivebox user data dir
+ARCHIVE_DIR = DATA_DIR / 'archive'                      # archivebox snapshot data dir
+
 
 try:
     JSONField = models.JSONField
@@ -17,14 +21,12 @@ except AttributeError:
 
 
 def forwards_func(apps, schema_editor):
-    from core.models import EXTRACTORS
-
     Snapshot = apps.get_model("core", "Snapshot")
     ArchiveResult = apps.get_model("core", "ArchiveResult")
 
     snapshots = Snapshot.objects.all()
     for snapshot in snapshots:
-        out_dir = Path(CONFIG['ARCHIVE_DIR']) / snapshot.timestamp
+        out_dir = ARCHIVE_DIR / snapshot.timestamp
 
         try:
             with open(out_dir / "index.json", "r") as f:
@@ -59,7 +61,7 @@ def forwards_func(apps, schema_editor):
 
 def verify_json_index_integrity(snapshot):
     results = snapshot.archiveresult_set.all()
-    out_dir = Path(CONFIG['ARCHIVE_DIR']) / snapshot.timestamp
+    out_dir = ARCHIVE_DIR / snapshot.timestamp
     with open(out_dir / "index.json", "r") as f:
         index = json.load(f)
 
