@@ -7,8 +7,9 @@ For more information on this file, see
 https://docs.djangoproject.com/en/2.1/howto/deployment/wsgi/
 """
 
+import archivebox                                       # noqa
+from archivebox.config.django import setup_django
 
-from archivebox.config import setup_django
 setup_django(in_memory_db=False, check_db=True)
 
 from django.core.wsgi import get_wsgi_application
