@@ -8,38 +8,35 @@ from pathlib import Path
 from datetime import datetime, timezone
 from typing import List, Optional, Iterator, Any, Union
 
+import abx
+
+from archivebox.config import VERSION, DATA_DIR, CONSTANTS
+from archivebox.config.common import SERVER_CONFIG, SHELL_CONFIG
+
 from .schema import Link
-from ..system import atomic_write
-from ..util import enforce_types
-from ..config import (
-    VERSION,
-    OUTPUT_DIR,
-    FOOTER_INFO,
-    DEPENDENCIES,
-    JSON_INDEX_FILENAME,
-    ARCHIVE_DIR_NAME,
-    ANSI
-)
+from archivebox.misc.system import atomic_write
+from archivebox.misc.util import enforce_types
 
 
-MAIN_INDEX_HEADER = {
-    'info': 'This is an index of site data archived by ArchiveBox: The self-hosted web archive.',
-    'schema': 'archivebox.index.json',
-    'copyright_info': FOOTER_INFO,
-    'meta': {
-        'project': 'ArchiveBox',
-        'version': VERSION,
-        'git_sha': VERSION,  # not used anymore, but kept for backwards compatibility
-        'website': 'https://ArchiveBox.io',
-        'docs': 'https://github.com/ArchiveBox/ArchiveBox/wiki',
-        'source': 'https://github.com/ArchiveBox/ArchiveBox',
-        'issues': 'https://github.com/ArchiveBox/ArchiveBox/issues',
-        'dependencies': DEPENDENCIES,
-    },
-}
 
 @enforce_types
-def generate_json_index_from_links(links: List[Link], with_headers: bool):
+def generate_json_index_from_links(links: List[Link], with_headers: bool=False):
+    MAIN_INDEX_HEADER = {
+        'info': 'This is an index of site data archived by ArchiveBox: The self-hosted web archive.',
+        'schema': 'archivebox.index.json',
+        'copyright_info': SERVER_CONFIG.FOOTER_INFO,
+        'meta': {
+            'project': 'ArchiveBox',
+            'version': VERSION,
+            'git_sha': VERSION,  # not used anymore, but kept for backwards compatibility
+            'website': 'https://ArchiveBox.io',
+            'docs': 'https://github.com/ArchiveBox/ArchiveBox/wiki',
+            'source': 'https://github.com/ArchiveBox/ArchiveBox',
+            'issues': 'https://github.com/ArchiveBox/ArchiveBox/issues',
+            'dependencies': abx.as_dict(abx.pm.hook.get_BINARIES()),
+        },
+    } if with_headers else {}
+    
     if with_headers:
         output = {
             **MAIN_INDEX_HEADER,
@@ -54,10 +51,10 @@ def generate_json_index_from_links(links: List[Link], with_headers: bool):
 
 
 @enforce_types
-def parse_json_main_index(out_dir: Path=OUTPUT_DIR) -> Iterator[Link]:
+def parse_json_main_index(out_dir: Path=DATA_DIR) -> Iterator[Link]:
     """parse an archive index json file and return the list of links"""
 
-    index_path = Path(out_dir) / JSON_INDEX_FILENAME
+    index_path = Path(out_dir) / CONSTANTS.JSON_INDEX_FILENAME
     if index_path.exists():
         with open(index_path, 'r', encoding='utf-8') as f:
             try:
@@ -68,7 +65,7 @@ def parse_json_main_index(out_dir: Path=OUTPUT_DIR) -> Iterator[Link]:
                 print("    {lightyellow}! Found an index.json in the project root but couldn't load links from it: {} {}".format(
                     err.__class__.__name__,
                     err,
-                    **ANSI,
+                    **SHELL_CONFIG.ANSI,
                 ))
                 return ()
 
@@ -77,14 +74,14 @@ def parse_json_main_index(out_dir: Path=OUTPUT_DIR) -> Iterator[Link]:
                     yield Link.from_json(link_json)
                 except KeyError:
                     try:
-                        detail_index_path = Path(OUTPUT_DIR) / ARCHIVE_DIR_NAME / link_json['timestamp']
+                        detail_index_path = CONSTANTS.ARCHIVE_DIR / link_json['timestamp']
                         yield parse_json_link_details(str(detail_index_path))
                     except KeyError: 
                         # as a last effort, try to guess the missing values out of existing ones
                         try:
                             yield Link.from_json(link_json, guess=True)
                         except KeyError:
-                            print("    {lightyellow}! Failed to load the index.json from {}".format(detail_index_path, **ANSI))
+                            # print("    {lightyellow}! Failed to load the index.json from {}".format(detail_index_path, **ANSI))
                             continue
     return ()
 
@@ -95,15 +92,16 @@ def write_json_link_details(link: Link, out_dir: Optional[str]=None) -> None:
     """write a json file with some info about the link"""
     
     out_dir = out_dir or link.link_dir
-    path = Path(out_dir) / JSON_INDEX_FILENAME
+    path = Path(out_dir) / CONSTANTS.JSON_INDEX_FILENAME
     atomic_write(str(path), link._asdict(extended=True))
 
 
 @enforce_types
-def parse_json_link_details(out_dir: Union[Path, str], guess: Optional[bool]=False) -> Optional[Link]:
+def parse_json_link_details(out_dir: Union[Path, str], guess: bool=False) -> Optional[Link]:
     """load the json link index from a given directory"""
-    existing_index = Path(out_dir) / JSON_INDEX_FILENAME
-    if existing_index.exists():
+    
+    existing_index = Path(out_dir) / CONSTANTS.JSON_INDEX_FILENAME
+    if os.access(existing_index, os.F_OK):
         with open(existing_index, 'r', encoding='utf-8') as f:
             try:
                 link_json = pyjson.load(f)
@@ -117,9 +115,10 @@ def parse_json_link_details(out_dir: Union[Path, str], guess: Optional[bool]=Fal
 def parse_json_links_details(out_dir: Union[Path, str]) -> Iterator[Link]:
     """read through all the archive data folders and return the parsed links"""
 
-    for entry in os.scandir(Path(out_dir) / ARCHIVE_DIR_NAME):
+
+    for entry in os.scandir(CONSTANTS.ARCHIVE_DIR):
         if entry.is_dir(follow_symlinks=True):
-            if (Path(entry.path) / 'index.json').exists():
+            if os.access((Path(entry.path) / 'index.json'), os.F_OK):
                 try:
                     link = parse_json_link_details(entry.path)
                 except KeyError:
@@ -138,13 +137,16 @@ class ExtendedEncoder(pyjson.JSONEncoder):
     """
 
     def default(self, obj):
-        cls_name = obj.__class__.__name__
+        cls_name = type(obj).__name__
 
         if hasattr(obj, '_asdict'):
             return obj._asdict()
 
         elif isinstance(obj, bytes):
             return obj.decode()
+        
+        elif isinstance(obj, Path):
+            return str(obj)
 
         elif isinstance(obj, datetime):
             return obj.isoformat()
@@ -153,12 +155,27 @@ class ExtendedEncoder(pyjson.JSONEncoder):
             return '{}: {}'.format(obj.__class__.__name__, obj)
 
         elif cls_name in ('dict_items', 'dict_keys', 'dict_values'):
-            return tuple(obj)
+            return list(obj)
+        
+        try:
+            return dict(obj)
+        except Exception:
+            pass
+        
+        try:
+            return list(obj)
+        except Exception:
+            pass
+        
+        try:
+            return str(obj)
+        except Exception:
+            pass
 
         return pyjson.JSONEncoder.default(self, obj)
 
 
 @enforce_types
-def to_json(obj: Any, indent: Optional[int]=4, sort_keys: bool=True, cls=ExtendedEncoder) -> str:
-    return pyjson.dumps(obj, indent=indent, sort_keys=sort_keys, cls=ExtendedEncoder)
+def to_json(obj: Any, indent: Optional[int]=4, sort_keys: bool=True, cls=ExtendedEncoder, default=None) -> str:
+    return pyjson.dumps(obj, indent=indent, sort_keys=sort_keys, cls=ExtendedEncoder, default=default)
 
